pyproject.toml
[project]
name = "wikimeda-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "kafka-python==2.0.2",
    "kafka>=1.3.5",
]
